test: wtf vs code python
