+++ .gitignore
@@ -1,0 +1,5 @@
+
+# Ignorar diretório de cache do Python
+__pycache__
+
+# Ignorar arquivo de banco de dados SQLite
