+++ functions/menuAnotacao.py
@@ -1,0 +1,12 @@
+while True:
+  def menu_anotacao():
+    print("=" *100)
+    print("")
+    print("Menu:")
+    print("1 - Inserir")
+    print("2 - Listar")
+    print("3 - Editar")
+    print("4 - Deletar")
+    print("5 - Sair")
+    print("")
+    print("=" *100)
